requirements.txt
aiosqlite
asyncpg
bcrypt
datasets
fastapi
fastapi-login
gunicorn
httpx
pandas
passlib
python-multipart
tortoise-orm
uvicorn
watchfiles
openai
python-dotenv